requirements.txt
Appium-Python-Client==4.0.0
pytest==8.3.2
allure-pytest==2.13.5
selene==2.0.0rc9
python-dotenv==1.0.1
requests==2.31.0
selenium==4.23.1